% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demo_estimates.R
\docType{data}
\name{attri_death_pm25}
\alias{attri_death_pm25}
\title{Demo dataset: Deaths attributable to PM.2.5 (Taiwan 2013)}
\format{
A data frame with 352 rows and  7 variables:
}
\source{
TBDC
}
\usage{
attri_death_pm25
}
\description{
A dataset containing the number of attributable deaths to PM2.5 exposure by townships. The values are simulated
from the original dataset.
}
\section{Variables}{
\itemize{
\item \code{town}: town code, MOI format (preferred)
\item \code{towncode_new}: town code, MOH format
\item \code{IHD_att_death}: IHD deaths attributable to PM2.5
\item \code{stroke_att_death}: stroke deaths attributable to PM2.5
\item \code{lung_att_death}: lung cancer deaths attributable to PM2.5
\item \code{copd_att_death}: COPD deaths attributable to PM2.5
\item \code{all_att_death}: Total deaths attributable to PM2.5
}
}

\seealso{
Other demo datasets: 
\code{\link{daly_B01}}
}
\concept{demo datasets}
\keyword{datasets}
