% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demo_estimates.R
\docType{data}
\name{daly_B01}
\alias{daly_B01}
\title{Demo dataset: DALY estimates of cancers (level-3)}
\format{
A data frame with 19554 rows and 8 variables:
}
\source{
TBDC
}
\usage{
daly_B01
}
\description{
A dataset containing the DALY, YLD and YLL of level-3 cancers (Taiwan 2015).
}
\section{Variables}{
\itemize{
\item \code{sex}: sex (0 = female and 1 = male)
\item \code{town}: town code, MOI format.
\item \code{year}: 2015
\item \code{cause_l3}: level-3 cause code
\item \code{daly_count}: DALY estimates (person-year)
\item \code{yld_count}: YLD estimates (person-year)
\item \code{yll_count}: YLL estimates (person-year)
\item \code{population}: population size
}
}

\seealso{
Other demo datasets: 
\code{\link{attri_death_pm25}}
}
\concept{demo datasets}
\keyword{datasets}
