% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionaries.R
\docType{data}
\name{dict_admin}
\alias{dict_admin}
\title{Reference dataset: dictionary for administrative codes/names}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 474 rows and 12 columns.
}
\source{
TBDC
}
\usage{
dict_admin
}
\description{
Reference dataset: dictionary for administrative codes/names
}
\section{Variables}{
\itemize{
\item \code{town_moh}/\code{city_moh}: townships and cities in MOH code format
\item \code{town_moi}/\code{city_moi}: townships and cities in MOI code format
\item \code{region}: region code
\item \code{country}: country code
\item \code{town_chn}/\code{city_chn}/\code{region_chn}/\code{country_chn}: chinse names of townships/cities/regions/country
\item \code{urb_2000}/\code{urb_2010}: urbanicity level in year 2000 and 2010.
}
}

\seealso{
Other reference datasets: 
\code{\link{TWpop_city}},
\code{\link{TWpop_region}},
\code{\link{TWpop_town}},
\code{\link{dict_cause}},
\code{\link{stand_pop}}
}
\concept{reference datasets}
\keyword{datasets}
