% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionaries.R
\docType{data}
\name{dict_cause}
\alias{dict_cause}
\title{Reference dataset: dictionary for cause outline and name}
\format{
An object of class \code{data.frame} with 712 rows and 7 columns.
}
\source{
TBDC
}
\usage{
dict_cause
}
\description{
Reference dataset: dictionary for cause outline and name
}
\section{Variables}{
\itemize{
\item \code{cause_id}: city code, MOI format
\item \code{cause_outline}: cause outline
\item \code{acause}: abbreviated cause name
\item \code{cause_name}: full cause name
\item \code{cause_short}: short cause name
\item \code{level}: level
\item \code{cause_name_ch}: full cause name in Chinese
}
}

\seealso{
Other reference datasets: 
\code{\link{TWpop_city}},
\code{\link{TWpop_region}},
\code{\link{TWpop_town}},
\code{\link{dict_admin}},
\code{\link{stand_pop}}
}
\concept{reference datasets}
\keyword{datasets}
