% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIfit.R
\name{HIfit}
\alias{HIfit}
\alias{plot.HIfit}
\title{Hierarchical Imputation Model}
\usage{
HIfit(X_fix, X_hier, Y, Ysd, Nsam, isDat, model = c("binomial", "normal"), ...)

\method{plot}{HIfit}(x)
}
\arguments{
\item{X_fix}{(numeric matrix) a model matrix (or design matrix) for fixed, non-hierarchical effects. We recommend using \link[stats:model.matrix]{stats::model.matrix} to generate this matrix.}

\item{X_hier}{(integer/character/factor matrix) a indicator matrix of hierarchical effects.}

\item{Y}{(numeric vector) an outcome vector. For \code{model = "binomial"}, this will be a vector of observed success probability (e.g. prevalence, incidence).
For \code{model = "normal"}, this will be a vector of observed sample means. \strong{Note that \code{NA}'s have to be replaced by any numeric values}.}

\item{Ysd}{(numeric vector) standard deviation for each observation. (required for \code{model = "normal"})}

\item{Nsam}{(integer vector) population, or sample size for each observation.}

\item{isDat}{(integer/logical vector) indicators which rows are observed (\code{1}, or \code{TRUE}) and which are missing values (\code{0}, or \code{FALSE}).}

\item{model}{(character) the model type, one of "binomial" or "normal".}

\item{...}{parameters for the sampling algorithm, such as
\itemize{
\item \code{chains}: number of the Markov chains (defaults to 4).
\item \code{cores}: number of cores to use when executing the chains in parallel (defaults to 1).
\item \code{iter}: number of iterations for each chain including warmup (defaults to 2000).
\item \code{warmup}: number of  warmup (aka burnin) iterations for each chain (defaults to \code{iter/2}).
\item \code{control }: a named list of parameters to control the sampler's behavior, such as
\itemize{
\item \code{adapt_delta}: between 0 and 1, defaults to 0.8.
\item \code{max_treedepth}: positive integer, defaults to 10.
\item ...: for other parameters, see the details in the documentation for the control argument in \code{\link[=rstan:stan]{rstan:stan()}}.
}
}}
}
\value{
A \code{HIfit} object containing the following components.
\itemize{
\item \code{fit}: an object of S4 class \link[rstan:stanfit-class]{rstan::stanfit}.
\item \code{dat}: a list of data passed \link[rstan:stanmodel-method-sampling]{rstan::sampling}.
\item \code{Y_obs}: a vector of observed outcomes. For \code{model = "binomial"}, this will equal to \code{Y/pop}. For For \code{model = "normal"}, this will equal to \code{Y}.
\item \code{Y_est}: a data.frame of predicted outcomes, corresponding to \code{obs}.
\item \code{model}: model family, one of "binomial" or "normal".
}
}
\description{
Fit a Bayesian Hierarchical model to make 'shrinkage' estimation and impute missing values in aggregated data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot the observations versus the predictions of HIfit
}}

