% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulated.R
\docType{data}
\name{multideath}
\alias{multideath}
\title{Simulated dataset: Death registry with multiple causes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 250000 rows and 13 columns.
}
\usage{
multideath
}
\description{
A dataset containing 250,000 records of multiple cause of death.
}
\section{Variables}{
\itemize{
\item \code{id}: identity key
\item \code{age}: age
\item \code{sex}: sex
\item \code{x1}-\code{x2}: true predictor variable (binary, can be NA)
\item \code{x3}-\code{x4}: noisy predictor variable (binary, can be NA)
\item \code{GUC}: main cause of death
\item \code{MC1}-\code{MC5}: supplemental cause of death
}
}

\keyword{datasets}
