% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_risk_factors.R
\docType{data}
\name{risk_3H}
\alias{risk_3H}
\title{Demo dataset: Estimates for 3H Risk factors  (blood pressure, glucose, LDL)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7315 rows and 9 columns.
}
\source{
TBDC
}
\usage{
risk_3H
}
\description{
Demo dataset: Estimates for 3H Risk factors  (blood pressure, glucose, LDL)
}
\section{Variables}{
\itemize{
\item \code{risk}: type of risk, one of "BP" (systolic blood pressure), "Glucose" (blood sugar), and "LDL" (low-density lipoprotein)
\item \code{age}: age group
\item \code{sex}: sex
\item \code{year}: year
\item \code{city}: city, MOI format (refer to \link{TWpop_city})
\item \code{region}: region (refer to \link{TWpop_region})
\item \code{mean_x}/\code{sd_x}: mean and standard deviation of risk exposure
\item \code{nsam}: \emph{ad hoc} sample size estimates (proportional to population, sum to ~7000)
}
}

\seealso{
Other risk factor estimates: 
\code{\link{risk_drink}},
\code{\link{risk_smoke}}
}
\concept{risk factor estimates}
\keyword{datasets}
