% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_risk_factors.R
\docType{data}
\name{risk_smoke}
\alias{risk_smoke}
\title{Demo dataset: Estimates for Risk factors  (Smoking)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 220 rows and 18 columns.
}
\source{
TBDC
}
\usage{
risk_smoke
}
\description{
Demo dataset: Estimates for Risk factors  (Smoking)
}
\section{Variables}{
\itemize{
\item \code{age}: age group
\item \code{sex}: sex
\item \code{year}: year
\item \code{city}: city, MOI format (refer to \link{TWpop_city})
\item \code{region}: region (refer to \link{TWpop_region})
\item \code{nsam}: no. of the samples surveyed
\item \code{csam}: no. of cases from the surveyed
\item \code{prev} (\code{prev_l}-\code{prev_u}): prevalence (95\% CI)
\item \code{mean_x}/\code{sd_x}: mean and standard deviation of cigarettes per day (日根)
\item \code{x_pth0}-\code{x_pth100}: minimum and maximum of cigarettes per day (日根)
\item \code{mean_z}/\code{sd_z}: mean and standard deviation of packs per year (年包)
\item \code{z_pth0}-\code{z_pth100}: minimum and maximum of packs per year (年包)
}
}

\seealso{
Other risk factor estimates: 
\code{\link{risk_3H}},
\code{\link{risk_drink}}
}
\concept{risk factor estimates}
\keyword{datasets}
