% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reference.R
\docType{data}
\name{stand_pop}
\alias{stand_pop}
\title{Reference dataset: age distribution of the standard population}
\format{
An object of class \code{data.frame} with 23 rows and 6 columns.
}
\source{
TBDC
}
\usage{
stand_pop
}
\description{
Reference dataset: age distribution of the standard population
}
\section{Variables}{
\itemize{
\item \code{age1}: age group in categorization 1
\item \code{percent1}: percentage in population of age group in categorization 1
\item \code{age2}: age group in categorization 2
\item \code{percent2}: percentage in population of age group in categorization 2
\item \code{age3}: age group in categorization 3
\item \code{percent3}: percentage in population of age group in categorization 3
}
}

\seealso{
Other reference datasets: 
\code{\link{TWpop_city}},
\code{\link{TWpop_region}},
\code{\link{TWpop_town}},
\code{\link{dict_admin}},
\code{\link{dict_cause}}
}
\concept{reference datasets}
\keyword{datasets}
