% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reference.R
\docType{data}
\name{TWpop_region}
\alias{TWpop_region}
\title{Reference dataset: Taiwan population by age, sex, year, region.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3240 rows and 5 columns.
}
\usage{
TWpop_region
}
\description{
Reference dataset: Taiwan population by age, sex, year, region.
}
\section{Variables}{
\itemize{
\item \code{region}: region code, MOI format
\item \code{age}: age in 5-year group
\item \code{sex}: sex
\item \code{year}: year
\item \code{population}: population size
}
}

\seealso{
Other reference datasets: 
\code{\link{TWpop_city}},
\code{\link{TWpop_town}},
\code{\link{dict_admin}},
\code{\link{dict_cause}},
\code{\link{stand_pop}}
}
\concept{reference datasets}
\keyword{datasets}
